requirements.txt
numpy
scipy
Pillow
matplotlib
scikit-image
tensorflow-cpu
h5py
graphviz
pydot
pydot_ng
termcolor
gym
tensorflow-probability
tensorflow-addons
tensorboard
keras-vis
tf-keras-vis